requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pyjwt
bcrypt
jinja2
python-multipart
markdown
pytest
gunicorn
firebase-admin
pydantic-settings
orjson